iterations need to be float or else all rmse will be 0

+++ src/k4a_por_calibration.cpp
@@ -266,7 +266,7 @@ bool K4APORCalibration::k4aAutoTuneExposure(const uint8_t target_blue_value, uin
 float K4APORCalibration::k4aRMSE(const float current, const float target, const int iteration, float &rmse)
 {
   float diff = current - target;
-  rmse = std::sqrt((1 / (iteration + 1)) * (diff * diff));
+  rmse = std::sqrt((1.0 / ((float)iteration + 1.0)) * (diff * diff));
   return rmse;
 }
